unix-socket: make unlink() optional in unix_stream_listen()

+++ unix-socket.c
@@ -108,7 +108,7 @@ fail:
 	return -1;
 }
 
-int unix_stream_listen(const char *path)
+int unix_stream_listen(const char *path, int replace)
 {
 	int fd, saved_errno;
 	struct sockaddr_un sa;
@@ -120,7 +120,8 @@ int unix_stream_listen(const char *path)
 	if (fd < 0)
 		goto fail;
 
-	unlink(path);
+	if (replace)
+		unlink(path);
 	if (bind(fd, (struct sockaddr *)&sa, sizeof(sa)) < 0)
 		goto fail;
